reflect: add REFLECT_MAX_FIELDS configuration variable, ise size_t instead of int

+++ tyndall/reflect/n_fields.h
@@ -8,11 +8,11 @@
 
 namespace nfields_detail
 {
-  template<int... I>
-  using index_sequence = std::integer_sequence<int, I...>;
+  template<size_t... I>
+  using index_sequence = std::integer_sequence<size_t, I...>;
 
-  template<int N>
-  using make_index_sequence = std::make_integer_sequence<int, N>;
+  template<size_t N>
+  using make_index_sequence = std::make_integer_sequence<size_t, N>;
 
   template <typename T>
   constexpr T unsafe_declval() noexcept {
@@ -38,35 +38,35 @@ namespace nfields_detail
     }
   };
 
-  template<typename T, int... I, typename = typename std::enable_if_t<std::is_copy_constructible_v<T>>>
+  template<typename T, size_t... I, typename = typename std::enable_if_t<std::is_copy_constructible_v<T>>>
   constexpr auto constructible(index_sequence<I...>) noexcept
     -> typename std::add_pointer_t<decltype(T{ ubiq_lref_constructor{I}... })>;
 
-  template<typename T, int... I, typename = typename std::enable_if_t<!std::is_copy_constructible_v<T>>>
+  template<typename T, size_t... I, typename = typename std::enable_if_t<!std::is_copy_constructible_v<T>>>
   constexpr auto constructible(index_sequence<I...>) noexcept
     -> typename std::add_pointer_t<decltype(T{ ubiq_rref_constructor{I}... })>;
 
-  template <class T, int N, typename = decltype(constructible<T>(make_index_sequence<N>()))>
-  using constructible_t = int;
+  template <class T, size_t N, typename = decltype(constructible<T>(make_index_sequence<N>()))>
+  using constructible_t = size_t;
 
   // n_fields_impl is overloaded with long and int to avoid multiple definitions
-  template<typename T, int N>
+  template<typename T, size_t N>
   constexpr auto n_fields_impl(long) noexcept -> constructible_t<T, N>
   {
     return N;
   }
 
-  template<typename T, int N>
-  constexpr int n_fields_impl(int) noexcept
+  template<typename T, size_t N>
+  constexpr size_t n_fields_impl(int) noexcept
   {
     return n_fields_impl<T, N-1>(0L);
   }
 }
 
 template<typename T>
-constexpr int n_fields() noexcept
+constexpr size_t n_fields() noexcept
 {
-  constexpr int max_field_count = sizeof(T)*CHAR_BIT;
+  constexpr size_t max_field_count = sizeof(T)*CHAR_BIT;
 
   using type = std::remove_cvref_t<T>;
